Fix logic for scaling watermarks
Removed unused variables, defines and files. Corrected logic in
predict() function to decide if it is time to start reclamation.
Corrected the name for file used to adjust watermarks and added
error messages if accessing this file results in error. Added
verbose flag. Fixed the logic for calling rescale_watermarks() and
computation of new watermark scale factor.

Signed-off-by: Khalid Aziz <khalid.aziz@oracle.com>