Fix zoneinfo parsing and maxwsf calculation
Fix parsing of /proc/zoneinfo file to allow for variation in
information layout across nodes. Fixed a possible arithmetic
exception issue with calculation of maximum watermark_scale_factor
value. Added computing current number of hugepages for future use.
Added suffix "msec" to a log message to the number of milliseconds
it was printing.

Signed-off-by: Khalid Aziz <khalid.aziz@oracle.com>