Fix conditional that triggers kswapd
The condition checked to see if kswapd should be started checks if
time taken for number of free pages to go below high watermark is
greater than time taken to reclaim enough pages to prevent that from
happening. This is inverted. Fix this conditional and use a
threshold greater than just the time it would take to reclaim enough
pages to ensure kswapd is not forced to wake up too often.

Signed-off-by: Khalid Aziz <khalid.aziz@oracle.com>