Fix crash on systems with less than 4G memory
When memoptimizer is started up on a machine with 4GB or less
memory, it encounters following exception and terminates:

"traps: memoptimizer[3025] trap divide error ip:402610 sp:7fffdf4a0f60
error:0 in memoptimizer[400000+5000]"

On a machine with less than 4GB of memory, all of the memory is
assigned to zone DMA32. memoptimizer looks at just the "Normal" zone
which has no pages assigned to it. As a result, in one of the
internal calculations when it divides by the number of managed
memory pages, it ends up dividing by 0 resulting in exception.
Update the code to consider memory in all zones except DMA and add
pages up across zones to get per node number of pages, and
watermarks.

Update version to 1.4.1 with this fix.

Signed-off-by: Khalid Aziz <khalid.aziz@oracle.com>