Add support for negative dentry limit tunable
Added detection and support for negative dentry limit system
tunable. There is a one time set up of value for negative dentry
limit at start up and this value is updated when number of hugepages
changes on the system.

Updated spec file with these changes and eventual release of
version 1.5.0

Signed-off-by: Khalid Aziz <khalid.aziz@oracle.com>
Reviewed-by: Laurence Rochfort <laurence.rochfort@oracle.com>