Change the name to adaptivemm
In order to identify the functionality of this memory optimizer, name
for the project is being updated to adaptivemm (Adaptive Memory
Manager). This commit updates source code and name in all of the
documentation as well as filenames. This is a major change and version
is being updated to "2.0.0" to reflect this.

Signed-off-by: Khalid Aziz <khalid.aziz@oracle.com>
Reviewed-by: Laurence Rochfort <laurence.rochfort@oracle.com>