adativemm: Fix when timestamp for saved data is calculated
Spec_before does not correctly represent the timestamp of the last call to
check_memory_pressure() as it is not static and its assignemnt is at the
start of check_memory_pressure(). Change spec_before to a static variable
so its value is retained between function calls and change its assignment
to the bottom of the check_memory_pressure() function. This allows for the
entry of the function having the before_time from the previous call to
check_memory_pressure().

Signed-off-by: Sidhartha Kumar <sidhartha.kumar@oracle.com>
Reviewed-by: Khalid Aziz <khalid.aziz@oracle.com>
Reviewed-by: Laurence Rochfort <laurence.rochfort@oracle.com>