adaptivemm: reorganize directory structure
This commit cleans up the directory structure of the adaptive tools.
The adaptived and adativemm projects are now in their own respective
directories. Further, the adativemm structure is also cleaned up to
seperate the code into a /src directory and adding the man file to a
/doc directory. The .spec file has been modified to support these changes.

Signed-off-by: Sidhartha Kumar <sidhartha.kumar@oracle.com>
Signed-off-by: Tom Hromatka <tom.hromatka@oracle.com>
TJH: Added links to the adaptived and adaptivemm READMEs.  Cleaned up
     then diff.sh and Github Actions logic.  Small whitespace changes.