adaptivemm: initial vm.vfs_cache_pressure tuning algorithm
Implement a simple vm.vfs_cache_pressure tuning algorithm. The algorithm
works by using a user supplied hint which is the PREFER_OBJECT_CACHING
option in the configuration file. This hint indicates if the user would
rather have their system preserve dentry and inode object caches at the
cost of higher memory utilizaiton. Some workloads are strongly impacted
by the level on dentry and inode caching so it may benefit these
workloads to have the mm system deprioritize reclaiming these caches. If
this option is selected, the inode and dentry numbers are monitored and
if they are projected to increase (via the slope of a best line fit),
the sysctl is lowered by 10% and if they are projected to decrease, the
sysctl is increased by 10%. The opposite behavior occurs if
PREFER_OBJECT_CACHING is specified to 0.

Signed-off-by: Sidhartha Kumar <sidhartha.kumar@oracle.com>